refactor and test
