+++ test/handlers/fill-event-processor.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import { DutchOrder } from '@uniswap/uniswapx-sdk'
 import { BigNumber } from 'ethers'
-import { FillEventProcessor } from '../../lib/handlers/check-order-status/service'
+import { FillEventProcessor } from '../../lib/handlers/check-order-status/fill-event-processor'
 import { FILL_EVENT_LOOKBACK_BLOCKS_ON, getSettledAmounts } from '../../lib/handlers/check-order-status/util'
 import { log } from '../../lib/Logging'
 import { MOCK_ENCODED_ORDER, MOCK_ORDER_ENTITY, MOCK_ORDER_HASH } from './test-data'
@@ -47,7 +47,37 @@ describe('processFillEvent', () => {
     }
   })
 
-  it('should processFillEvent', async () => {
+  it('should return the settled amount', async () => {
+    const parsedOrder = DutchOrder.parse(MOCK_ENCODED_ORDER, 1)
+    const quoteId = '1'
+    const order = MOCK_ORDER_ENTITY
+    const chainId = 1
+    const startingBlockNumber = 0
+
+    const getSettledAmountsReturn = {
+      tokenOut: 'out',
+      amountOut: '1',
+      tokenIn: 'in',
+      amountIn: '2',
+    }
+
+    getSettledAmountsMock.mockReturnValue([getSettledAmountsReturn])
+
+    let fillEventProcessor = new FillEventProcessor(FILL_EVENT_LOOKBACK_BLOCKS_ON)
+
+    const response = await fillEventProcessor.processFillEvent({
+      provider: providerMock as any,
+      fillEvent: MOCK_FILL_EVENT,
+      parsedOrder,
+      quoteId,
+      order,
+      chainId,
+      startingBlockNumber,
+    })
+    expect(response).toEqual([getSettledAmountsReturn])
+  })
+
+  it('should return the settled amount', async () => {
     const parsedOrder = DutchOrder.parse(MOCK_ENCODED_ORDER, 1)
     const quoteId = '1'
     const order = MOCK_ORDER_ENTITY
